<Add> test user successfully <Next> items test and other
