+++ src/test/java/ru/practicum/shareit/items/ItemServiceTest.java
@@ -5,172 +5,226 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.test.annotation.DirtiesContext;
+import ru.practicum.shareit.booking.BookingInputDto;
+import ru.practicum.shareit.booking.BookingService;
 import ru.practicum.shareit.exception.NotFoundException;
 import ru.practicum.shareit.exception.ValidatorException;
 import ru.practicum.shareit.item.ItemDto;
 import ru.practicum.shareit.item.ItemInputDto;
 import ru.practicum.shareit.item.ItemService;
-import ru.practicum.shareit.item.comment.CommentDto;
 import ru.practicum.shareit.item.comment.CommentInputDto;
+import ru.practicum.shareit.request.ItemRequestDto;
+import ru.practicum.shareit.request.ItemRequestInputDto;
+import ru.practicum.shareit.request.ItemRequestService;
 import ru.practicum.shareit.user.UserDto;
+import ru.practicum.shareit.user.UserService;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.equalTo;
-import static org.hamcrest.Matchers.notNullValue;
 import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.ArgumentMatchers.*;
-import static org.mockito.Mockito.when;
-import static ru.practicum.shareit.user.UserMapper.dtoToUser;
-
-/**
- * @author MR.k0F31n
- */
 
 @SpringBootTest
 @AutoConfigureTestDatabase
 @RequiredArgsConstructor(onConstructor_ = @Autowired)
 public class ItemServiceTest {
-    @MockBean
-    private final ItemService service;
+    private final UserService service;
+    private final ItemService items;
+    private final ItemRequestService itemRequestService;
+    private final BookingService bookingService;
 
-    private final UserDto user1 = new UserDto(1L, "user1", "user1@mail.com");
-    private final UserDto user2 = new UserDto(2L, "user2", "user2@mail.com");
-    private final ItemInputDto item1 = new ItemInputDto("item1", "descr1", true, null);
-    private final CommentDto commentDto1 = new CommentDto(1L, "text from comment1", 1L, user2.getName(),
-            LocalDateTime.now());
-    private final ItemDto itemDto2 = new ItemDto(2L, "item2", "descr2", true,
-            dtoToUser(user1), null, null, List.of(commentDto1), null);
-    private final ItemDto itemDto1 = new ItemDto(1L, "item1", "descr1", true,
-            dtoToUser(user1), null, null, new ArrayList<>(), null);
+    private final UserDto user1 = new UserDto(1L, "name1", "emai1@mail.com");
+    private final ItemInputDto inputItem = new ItemInputDto("Отвертка", "Отвертка в печень ни один тест не вечен",
+            true, null);
+    private final BookingInputDto inputDto = new BookingInputDto(LocalDateTime.now().plusSeconds(20),
+            LocalDateTime.now().plusMinutes(2), 1L);
 
+
+    @DirtiesContext
     @Test
-    void testGetAllItemsByOwnerId_returnListItemDto_length2() {
-        when(service.getAllItemsByOwner(anyLong(), anyInt(), anyInt())).thenReturn(List.of(itemDto1, itemDto2));
+    void createItemFromRequest_returnItemWitchRequest() {
+        UserDto user = service.createNewUser(new UserDto(null, "name", "mail@mail.ru"));
+        ItemRequestDto request = itemRequestService.createNewItemRequest(new ItemRequestInputDto("need item"),
+                user.getId());
+        ItemDto item = items.createNewItem(new ItemInputDto("Item", "descr", true, request.getId()),
+                user.getId());
 
-        List<ItemDto> items = service.getAllItemsByOwner(user1.getId(), 0, 10);
-
-        assertNotNull(items, "Предметов нет");
-        assertEquals(2, items.size(), "Количесвто предметов не совпадает");
+        assertEquals(user.getId(), item.getRequestId());
     }
 
+    @DirtiesContext
     @Test
-    void testCreateNewUser_returnItemDto() {
-        when(service.createNewItem(any(ItemInputDto.class), anyLong())).thenReturn(itemDto1);
-
-        ItemDto item = service.createNewItem(item1, user1.getId());
-
-        assertThat(item.getId(), notNullValue());
-        assertThat(item.getName(), equalTo(itemDto1.getName()));
-        assertThat(item.getDescription(), equalTo(itemDto1.getDescription()));
+    void createItemFromRequest_requestNotFound() {
+        UserDto user = service.createNewUser(new UserDto(null, "name", "mail@mail.ru"));
+        ItemRequestDto request = itemRequestService.createNewItemRequest(new ItemRequestInputDto("need item"),
+                user.getId());
+        assertThrows(NotFoundException.class, () -> items.createNewItem(new ItemInputDto("Item", "descr", true, 999L),
+                user.getId()));
     }
 
+    @DirtiesContext
     @Test
-    void updateItem_correctUpdate() {
-        ItemDto updateItem = itemDto1;
-        updateItem.setAvailable(false);
-        updateItem.setDescription("Update");
-        updateItem.setName("Update");
+    void updateUser_inputNull_returnNull() {
+        UserDto updateUser = user1;
+        updateUser.setId(1L);
+        updateUser.setName(null);
+        updateUser.setEmail(null);
 
-        when(service.updateItem(any(ItemDto.class), anyLong(), anyLong())).thenReturn(updateItem);
-
-        ItemDto itemBeforeUpdate = service.updateItem(updateItem, itemDto1.getId(), user1.getId());
-
-        assertEquals(updateItem.getName(), itemBeforeUpdate.getName(), "Name not update");
-        assertEquals(updateItem.getAvailable(), itemBeforeUpdate.getAvailable(), "Available not update");
-        assertEquals(updateItem.getDescription(), itemBeforeUpdate.getDescription(), "Description not update");
+        assertNull(service.updateUser(updateUser, 1L), "Обновились в null!");
     }
 
+    @DirtiesContext
     @Test
-    void updateItem_allNull_expectedNull() {
-        ItemDto updateItem = itemDto1;
-        updateItem.setAvailable(null);
-        updateItem.setDescription(null);
-        updateItem.setName(null);
+    void updateItem_AllUpdate_returnDto() {
+        service.createNewUser(user1);
+        ItemDto item1 = items.createNewItem(inputItem, user1.getId());
+        item1.setName("ОтВерТка апргрейд");
 
-        when(service.updateItem(any(ItemDto.class), anyLong(), anyLong())).thenReturn(null);
+        ItemDto itemBeforeUpdateName = items.updateItem(item1, item1.getId(), user1.getId());
 
-        ItemDto itemBeforeUpdate = service.updateItem(updateItem, itemDto1.getId(), user1.getId());
+        assertEquals(item1.getName(), itemBeforeUpdateName.getName());
 
-        assertNull(itemBeforeUpdate);
+        item1.setDescription("Убивает на растоянии");
+
+        ItemDto itemBeforeUpdateDesc = items.updateItem(item1, item1.getId(), user1.getId());
+
+        assertEquals(item1.getName(), itemBeforeUpdateDesc.getName());
+        assertEquals(item1.getDescription(), itemBeforeUpdateDesc.getDescription());
+
+        item1.setAvailable(false);
+
+        ItemDto itemBeforeUpdateAvl = items.updateItem(item1, item1.getId(), user1.getId());
+
+        assertEquals(item1.getName(), itemBeforeUpdateAvl.getName());
+        assertEquals(item1.getDescription(), itemBeforeUpdateAvl.getDescription());
+        assertEquals(item1.getAvailable(), itemBeforeUpdateAvl.getAvailable());
     }
 
-    @Test
-    void updateItem_availableIsNull_availableNotUpdate() {
-        ItemDto updateItem = itemDto1;
-        updateItem.setAvailable(null);
-
-        when(service.updateItem(any(ItemDto.class), anyLong(), anyLong())).thenReturn(itemDto1);
-
-        ItemDto itemBeforeUpdate = service.updateItem(updateItem, itemDto1.getId(), user1.getId());
-
-        assertEquals(itemDto1.getAvailable(), itemBeforeUpdate.getAvailable());
-    }
-
-    @Test
-    void updateItem_descriptionIsNull_descriptionNotUpdate() {
-        ItemDto updateItem = itemDto1;
-        updateItem.setDescription(null);
-
-        when(service.updateItem(any(ItemDto.class), anyLong(), anyLong())).thenReturn(itemDto1);
-
-        ItemDto itemBeforeUpdate = service.updateItem(updateItem, itemDto1.getId(), user1.getId());
-
-        assertEquals(itemDto1.getDescription(), itemBeforeUpdate.getDescription());
-    }
-
+    @DirtiesContext
     @Test
     void updateItem_userNotOwner_expectedError() {
-        ItemDto updateItem = itemDto1;
-        updateItem.setName("Name update");
+        service.createNewUser(user1);
+        ItemDto item1 = items.createNewItem(inputItem, user1.getId());
+        item1.setName("ОтВерТка апргрейд");
 
-        when(service.updateItem(any(ItemDto.class), anyLong(), anyLong())).thenThrow(NotFoundException.class);
-
-        assertThrows(NotFoundException.class, () -> service.updateItem(updateItem, itemDto1.getId(), user2.getId()));
+        assertThrows(NotFoundException.class, () -> items.updateItem(item1, item1.getId(), 2L));
     }
 
+    @DirtiesContext
     @Test
-    void getItemById_returnItemDto() {
-        when(service.getItemDtoById(anyLong(), anyLong())).thenReturn(itemDto1);
-
-        ItemDto findItem = service.getItemDtoById(1L, 1L);
-
-        assertEquals(itemDto1.getName(), findItem.getName());
+    void getItem_itemNotFound_userExist() {
+        assertThrows(NotFoundException.class, () -> items.getItemDtoById(99L, 1L));
     }
 
+    @DirtiesContext
     @Test
-    void getItemByID_itemNotFound_expectedError() {
-        when(service.getItemDtoById(anyLong(), anyLong())).thenThrow(NotFoundException.class);
-
-        assertThrows(NotFoundException.class, () -> service.getItemDtoById(99L, 1L));
+    void getItem_itemNotFound_userNotExist() {
+        assertThrows(NotFoundException.class, () -> items.getItemDtoById(99L, 2L));
     }
 
+    @DirtiesContext
     @Test
-    void searchItem_returnListItemDto_length1() {
-        when(service.searchItem(anyString(), anyInt(), anyInt())).thenReturn(List.of(itemDto1));
-
-        assertEquals(1, service.searchItem("iTem1", 0, 10).size());
+    void getItem_returnDto() {
+        final UserDto userOne = new UserDto(1L, "name1", "emai2313@mail.com");
+        UserDto userDto = service.createNewUser(userOne);
+        ItemDto item1 = items.createNewItem(inputItem, userDto.getId());
+        ItemDto findItem = items.getItemDtoById(item1.getId(), userDto.getId());
+        assertEquals(item1.getName(), findItem.getName());
+        assertEquals(item1.getDescription(), findItem.getDescription());
+        items.deleteItem(item1.getId(), userDto.getId());
     }
 
+    @DirtiesContext
     @Test
-    void addComment_returnCommentDto() {
-        when(service.addComment(anyLong(), anyLong(), any(CommentInputDto.class))).thenReturn(commentDto1);
-        CommentInputDto comment = new CommentInputDto();
-        comment.setText("text from comment1");
+    void getItemNotOwner_returnDto() {
+        UserDto user = service.createNewUser(user1);
+        UserDto user2 = new UserDto(2L, "name2", "emai2@mail.com");
+        service.createNewUser(user2);
 
-        CommentDto commentDto = service.addComment(2L, 1L, comment);
+        ItemDto item1 = items.createNewItem(inputItem, user.getId());
+        ItemDto findItem = items.getItemDtoById(item1.getId(), user2.getId());
 
-        assertEquals(commentDto1.getText(), commentDto.getText());
+        assertEquals(item1.getName(), findItem.getName());
+        assertEquals(item1.getDescription(), findItem.getDescription());
+
+        items.deleteItem(item1.getId(), user.getId());
+        service.deleteUserById(user2.getId());
     }
 
+    @DirtiesContext
     @Test
-    void addComment_commentIsNull_expectedError() {
-        when(service.addComment(anyLong(), anyLong(), any(CommentInputDto.class))).thenThrow(ValidatorException.class);
+    void searchItem_returnDto() {
+        service.createNewUser(user1);
+        ItemDto item1 = items.createNewItem(inputItem, user1.getId());
+        List<ItemDto> findItem = items.searchItem("ОтВЕРТка", 0, 10);
+        assertEquals(item1.getName(), findItem.get(0).getName());
+        assertEquals(item1.getDescription(), findItem.get(0).getDescription());
+        items.deleteItem(item1.getId(), user1.getId());
+    }
 
-        assertThrows(ValidatorException.class, () -> service.addComment(2L, 1L, new CommentInputDto()));
+    @DirtiesContext
+    @Test
+    void searchItem_emptyRequest_emptyList() {
+        service.createNewUser(user1);
+        ItemDto item1 = items.createNewItem(inputItem, user1.getId());
+        List<ItemDto> findItem = items.searchItem("", 0, 10);
+        assertEquals(0, findItem.size());
+        items.deleteItem(item1.getId(), user1.getId());
+    }
+
+    @DirtiesContext
+    @Test
+    void searchItem_emptySpace_emptyList() {
+        service.createNewUser(user1);
+        ItemDto item1 = items.createNewItem(inputItem, user1.getId());
+        List<ItemDto> findItem = items.searchItem("  ", 0, 10);
+        assertEquals(0, findItem.size());
+        items.deleteItem(item1.getId(), user1.getId());
+    }
+
+    @DirtiesContext
+    @Test
+    void addComment_textIsEmpty() {
+        CommentInputDto input = new CommentInputDto();
+        input.setText(" ");
+        assertThrows(ValidatorException.class, () -> items.addComment(1L, 1L, input));
+    }
+
+    @DirtiesContext
+    @Test
+    void addComment_userNotRentItem() {
+        CommentInputDto input = new CommentInputDto();
+        input.setText("User not rent Item");
+        assertThrows(ValidatorException.class, () -> items.addComment(1L, 1L, input));
+    }
+
+    @DirtiesContext
+    @Test
+    void deleteItemItem() {
+        UserDto userDto = service.createNewUser(user1);
+        ItemDto itemFromDelete = items.createNewItem(new ItemInputDto("ItemFromDelete", "from delete desc", true, null), userDto.getId());
+        items.deleteItem(itemFromDelete.getId(), 1L);
+    }
+
+    @DirtiesContext
+    @Test
+    void deleteItemItem_userNotFound_expectedError() {
+        assertThrows(NotFoundException.class, () -> items.deleteItem(1L, 999L));
+    }
+
+    @DirtiesContext
+    @Test
+    void deleteItemItem_itemNotFound_expectedError() {
+        assertThrows(NotFoundException.class, () -> items.deleteItem(999L, 1L));
+    }
+
+    @DirtiesContext
+    @Test
+    void deleteItemItem_userNotOwner_expectedError() {
+        UserDto userDto = service.createNewUser(user1);
+        UserDto userDto1 = service.createNewUser(new UserDto(2L, "kto-to", "ctoTo@mail.com"));
+        ItemDto itemFromDelete = items.createNewItem(new ItemInputDto("ItemFromDelete", "from delete desc", true, null), userDto.getId());
+        assertThrows(NotFoundException.class, () -> items.deleteItem(itemFromDelete.getId(), userDto1.getId()));
     }
 }
